Feature/events

+++ go-event-ticket-service/internal/events/infrastructure/params/common.params.go
@@ -1,6 +1,6 @@
 package params
 
 type PaginateParams struct {
-	Limit  int8
-	Offset int8
+	Limit  int
+	Offset int
 }
